Made current activity implied by the start time of the activity queue. Gave villagers an activity queue. Fixed the label field of the Time table.

+++ VillageOfFate.Services/DALServices/VillagerActivityService.cs
@@ -7,11 +7,13 @@ namespace VillageOfFate.Services.DALServices;
 public class VillagerActivityService(DataContext context, TimeService time) {
 	public async Task PopAsync(VillagerDto villager) {
 		villager = context.Villagers.Entry(villager).Entity;
+		if (villager.CurrentActivity == null) return;
 		context.Activities.Remove(villager.CurrentActivity);
 
-		var newActivity = villager.ActivityQueue.First();
-		villager.CurrentActivity = newActivity;
-		newActivity.StartTime = await time.GetAsync(TimeLabel.World);
+		var newActivity = villager.ActivityQueue.FirstOrDefault();
+		if (newActivity != null) {
+			newActivity.StartTime = await time.GetAsync(TimeLabel.World);
+		}
 
 		await context.SaveChangesAsync();
 	}
